refactor(web_fetch): extract content and result modules
