+++ lib/jido_browser/web_fetch.ex
@@ -10,8 +10,10 @@ defmodule Jido.Browser.WebFetch do
 
   alias Jido.Browser.Error
   alias Jido.Browser.WebFetch.Cache
+  alias Jido.Browser.WebFetch.Content
   alias Jido.Browser.WebFetch.DestinationPolicy
   alias Jido.Browser.WebFetch.Options
+  alias Jido.Browser.WebFetch.Result
   alias Jido.Browser.WebFetch.URLRules
 
   @redirect_statuses [301, 302, 303, 307, 308]
@@ -26,66 +28,6 @@ defmodule Jido.Browser.WebFetch do
   @post_redirect_body_options [:body, :json, :form, :form_multipart]
   @body_headers ["content-length", "content-type"]
   @cross_origin_headers ["authorization", "proxy-authorization", "cookie", "cookie2"]
-  @html_content_types ["text/html", "application/xhtml+xml"]
-  @text_content_types [
-    "text/plain",
-    "text/markdown",
-    "text/csv",
-    "text/xml",
-    "application/xml",
-    "application/json",
-    "application/ld+json"
-  ]
-  @document_content_types %{
-    "application/pdf" => :pdf,
-    "application/msword" => :word_processing,
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document" => :word_processing,
-    "application/vnd.ms-word.document.macroenabled.12" => :word_processing,
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.template" => :word_processing,
-    "application/vnd.ms-word.template.macroenabled.12" => :word_processing,
-    "application/vnd.ms-excel" => :spreadsheet,
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" => :spreadsheet,
-    "application/vnd.ms-excel.sheet.macroenabled.12" => :spreadsheet,
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.template" => :spreadsheet,
-    "application/vnd.ms-excel.template.macroenabled.12" => :spreadsheet,
-    "application/vnd.ms-powerpoint" => :presentation,
-    "application/vnd.openxmlformats-officedocument.presentationml.presentation" => :presentation,
-    "application/vnd.ms-powerpoint.presentation.macroenabled.12" => :presentation,
-    "application/vnd.openxmlformats-officedocument.presentationml.slideshow" => :presentation,
-    "application/vnd.openxmlformats-officedocument.presentationml.template" => :presentation,
-    "application/vnd.oasis.opendocument.text" => :word_processing,
-    "application/vnd.oasis.opendocument.spreadsheet" => :spreadsheet,
-    "application/vnd.oasis.opendocument.presentation" => :presentation,
-    "application/rtf" => :word_processing,
-    "text/rtf" => :word_processing,
-    "application/epub+zip" => :ebook,
-    "message/rfc822" => :email,
-    "application/vnd.ms-outlook" => :email
-  }
-  @document_extensions %{
-    "pdf" => :pdf,
-    "doc" => :word_processing,
-    "docx" => :word_processing,
-    "docm" => :word_processing,
-    "dotx" => :word_processing,
-    "dotm" => :word_processing,
-    "odt" => :word_processing,
-    "rtf" => :word_processing,
-    "xls" => :spreadsheet,
-    "xlsx" => :spreadsheet,
-    "xlsm" => :spreadsheet,
-    "xlsb" => :spreadsheet,
-    "ods" => :spreadsheet,
-    "ppt" => :presentation,
-    "pptx" => :presentation,
-    "pptm" => :presentation,
-    "ppsx" => :presentation,
-    "odp" => :presentation,
-    "epub" => :ebook,
-    "eml" => :email,
-    "msg" => :email
-  }
-
   @type result :: %{
           required(:url) => String.t(),
           required(:final_url) => String.t(),
@@ -160,7 +102,8 @@ defmodule Jido.Browser.WebFetch do
       try do
         with {:ok, response, final_url, _final_uri} <- fetch_with_redirects(url, request_opts),
              :ok <- validate_http_status(response, url),
-             {:ok, result} <- build_result(url, final_url, response, request_opts) do
+             {:ok, content} <- Content.extract(final_url, response, request_opts),
+             {:ok, result} <- Result.build(url, final_url, content, request_opts) do
           Cache.store(url, request_opts, result)
           {:ok, result}
         end
@@ -303,7 +246,7 @@ defmodule Jido.Browser.WebFetch do
 
   defp redirect_location(response) do
     response.headers
-    |> response_header("location")
+    |> Content.response_header("location")
     |> List.first()
     |> case do
       location when is_binary(location) ->
@@ -423,159 +366,6 @@ defmodule Jido.Browser.WebFetch do
     {:ok, Map.put(response, :final_url, final_url), final_url, final_uri}
   end
 
-  defp build_result(url, final_url, response, opts) do
-    content_type = response_content_type(response)
-    document_type = extractable_document_type(content_type, final_url, response.body)
-
-    cond do
-      content_type in @html_content_types ->
-        build_html_result(url, final_url, response.body, content_type, opts)
-
-      not is_nil(document_type) ->
-        build_document_result(url, final_url, response.body, content_type, document_type, opts)
-
-      text_content_type?(content_type) ->
-        build_text_result(url, final_url, response.body, content_type, opts)
-
-      true ->
-        {:error,
-         Error.adapter_error("Unsupported content type for web fetch", %{
-           error_code: :unsupported_content_type,
-           content_type: content_type
-         })}
-    end
-  end
-
-  defp build_html_result(url, final_url, body, content_type, opts) when is_binary(body) do
-    selector = opts[:selector]
-
-    with {:ok, document} <- parse_document(body),
-         {:ok, html} <- select_html(document, body, selector),
-         {:ok, title} <- extract_title(document),
-         {:ok, content} <- format_html(html, opts[:format], opts) do
-      finalize_result(url, final_url, content, title, content_type, :html, opts)
-    end
-  end
-
-  defp build_html_result(_url, _final_url, body, content_type, _opts) do
-    {:error,
-     Error.adapter_error("Unexpected response body for HTML fetch", %{
-       error_code: :unavailable,
-       content_type: content_type,
-       body: body
-     })}
-  end
-
-  defp build_text_result(url, final_url, body, content_type, opts) when is_binary(body) do
-    with :ok <- validate_non_html_options(content_type, opts),
-         {:ok, content} <- format_text(body, opts[:format]) do
-      finalize_result(url, final_url, content, nil, content_type, :text, opts)
-    end
-  end
-
-  defp build_text_result(_url, _final_url, body, content_type, _opts) do
-    {:error,
-     Error.adapter_error("Unexpected response body for text fetch", %{
-       error_code: :unavailable,
-       content_type: content_type,
-       body: body
-     })}
-  end
-
-  defp build_document_result(url, final_url, body, content_type, document_type, opts) when is_binary(body) do
-    with :ok <- validate_non_html_options(content_type, opts),
-         {:ok, text, metadata} <- extract_document_content(body, final_url, content_type, document_type, opts) do
-      finalize_result(
-        url,
-        final_url,
-        text,
-        document_title(metadata, final_url),
-        content_type,
-        document_type,
-        opts,
-        metadata
-      )
-    end
-  end
-
-  defp build_document_result(_url, _final_url, body, content_type, _document_type, _opts) do
-    {:error,
-     Error.adapter_error("Unexpected response body for document fetch", %{
-       error_code: :unavailable,
-       content_type: content_type,
-       body: body
-     })}
-  end
-
-  defp build_response(opts, attrs) do
-    passages = maybe_build_passages(attrs.content, attrs.title, attrs.final_url, opts[:citations])
-
-    %{
-      url: attrs.url,
-      final_url: attrs.final_url,
-      title: attrs.title,
-      content: attrs.content,
-      format: opts[:format],
-      content_type: attrs.content_type,
-      document_type: attrs.document_type,
-      retrieved_at: retrieved_at(),
-      estimated_tokens: estimate_tokens(attrs.content),
-      original_estimated_tokens: attrs.original_estimated_tokens,
-      truncated: attrs.truncated,
-      filtered: attrs.filtered,
-      focus_matches: attrs.focus_matches,
-      cached: false,
-      citations: %{enabled: opts[:citations]},
-      passages: passages
-    }
-    |> maybe_put_metadata(attrs.metadata)
-  end
-
-  defp finalize_result(url, final_url, content, title, content_type, document_type, opts, metadata \\ nil) do
-    with {:ok, filtered_content, filtered, focus_matches} <- maybe_filter_content(content, opts),
-         {final_content, truncated, original_estimated_tokens} <-
-           maybe_truncate(filtered_content, opts[:max_content_tokens]) do
-      attrs = %{
-        url: url,
-        final_url: final_url,
-        content: final_content,
-        title: title,
-        content_type: content_type,
-        document_type: document_type,
-        truncated: truncated,
-        filtered: filtered,
-        focus_matches: focus_matches,
-        original_estimated_tokens: original_estimated_tokens,
-        metadata: metadata
-      }
-
-      {:ok, build_response(opts, attrs)}
-    end
-  end
-
-  defp validate_non_html_options(content_type, opts) do
-    cond do
-      opts[:selector] ->
-        {:error,
-         Error.invalid_error("Selector filtering is only supported for HTML content", %{
-           error_code: :invalid_input,
-           selector: opts[:selector],
-           content_type: content_type
-         })}
-
-      opts[:format] == :html ->
-        {:error,
-         Error.invalid_error("HTML output is only supported for HTML content", %{
-           error_code: :invalid_input,
-           format: :html,
-           content_type: content_type
-         })}
-
-      true ->
-        :ok
-    end
-  end
-
   defp fetch_prepared_destination(backend, url, opts) do
     case Keyword.pop(opts, :destination_addresses) do
       {nil, request_opts} -> backend.fetch(url, request_opts)
@@ -623,340 +413,5 @@ defmodule Jido.Browser.WebFetch do
        status: status,
        url: url
      })}
-  end
-
-  defp parse_document(body) do
-    case Floki.parse_document(body) do
-      {:ok, document} ->
-        {:ok, document}
-
-      {:error, reason} ->
-        {:error, Error.adapter_error("Failed to parse fetched HTML", %{error_code: :unavailable, reason: reason})}
-    end
-  end
-
-  defp select_html(_document, body, nil), do: {:ok, body}
-  defp select_html(document, _body, ""), do: select_html(document, nil, nil)
-
-  defp select_html(document, _body, selector) do
-    nodes = Floki.find(document, selector)
-
-    if nodes == [] do
-      {:error,
-       Error.invalid_error("Selector did not match any elements in fetched HTML", %{
-         error_code: :invalid_input,
-         selector: selector
-       })}
-    else
-      {:ok, Floki.raw_html(nodes)}
-    end
-  end
-
-  defp extract_title(document) do
-    title =
-      document
-      |> Floki.find("title")
-      |> Floki.text(sep: " ")
-      |> String.trim()
-      |> blank_to_nil()
-
-    {:ok, title}
-  end
-
-  defp format_html(html, :html, _opts), do: {:ok, html}
-
-  defp format_html(html, :text, _opts) do
-    with {:ok, fragment} <- parse_fragment(html) do
-      {:ok, fragment |> Floki.text(sep: "\n") |> String.trim()}
-    end
-  end
-
-  defp format_html(html, :markdown, _opts) do
-    {:ok, Html2Markdown.convert(html) |> String.trim()}
-  rescue
-    error ->
-      {:error,
-       Error.adapter_error("Failed to convert fetched HTML to markdown", %{error_code: :unavailable, reason: error})}
-  end
-
-  defp format_text(text, :text), do: {:ok, String.trim(text)}
-  defp format_text(text, :markdown), do: {:ok, String.trim(text)}
-
-  defp format_text(_text, :html) do
-    {:error,
-     Error.invalid_error("HTML output is only supported for HTML content", %{
-       error_code: :invalid_input
-     })}
-  end
-
-  defp parse_fragment(html) do
-    case Floki.parse_fragment(html) do
-      {:ok, fragment} ->
-        {:ok, fragment}
-
-      {:error, reason} ->
-        {:error,
-         Error.adapter_error("Failed to parse fetched HTML fragment", %{error_code: :unavailable, reason: reason})}
-    end
-  end
-
-  defp maybe_filter_content(content, opts) do
-    case opts[:focus_terms] do
-      [] ->
-        {:ok, content, false, 0}
-
-      terms ->
-        sections = split_sections(content)
-        matching_indexes = matching_section_indexes(sections, terms)
-        window = max(opts[:focus_window] || 0, 0)
-        kept_indexes = expand_focus_window(matching_indexes, window, length(sections))
-        filtered_content = render_section_slice(sections, kept_indexes)
-
-        {:ok, filtered_content, true, length(matching_indexes)}
-    end
-  end
-
-  defp maybe_truncate(content, nil), do: {content, false, estimate_tokens(content)}
-
-  defp maybe_truncate(content, max_content_tokens) when is_integer(max_content_tokens) and max_content_tokens > 0 do
-    original_estimated_tokens = estimate_tokens(content)
-
-    if original_estimated_tokens <= max_content_tokens do
-      {content, false, original_estimated_tokens}
-    else
-      char_limit = max_content_tokens * 4
-      truncated = String.slice(content, 0, char_limit) |> String.trim()
-      {truncated, true, original_estimated_tokens}
-    end
-  end
-
-  defp maybe_truncate(content, _other), do: {content, false, estimate_tokens(content)}
-
-  defp maybe_build_passages(_content, _title, _url, false), do: []
-
-  defp maybe_build_passages(content, title, url, true) do
-    content
-    |> split_sections()
-    |> Enum.reject(&(&1 == ""))
-    |> Enum.reduce({[], 0, 0}, fn section, {passages, cursor, index} ->
-      start_char = cursor
-      end_char = start_char + String.length(section)
-
-      passage = %{
-        index: index,
-        start_char: start_char,
-        end_char: end_char,
-        text: section,
-        title: title,
-        url: url
-      }
-
-      {[passage | passages], end_char + 2, index + 1}
-    end)
-    |> elem(0)
-    |> Enum.reverse()
-    |> Enum.take(50)
-  end
-
-  defp split_sections(content) do
-    content
-    |> String.split(~r/\n\s*\n+/, trim: true)
-    |> case do
-      [] -> [String.trim(content)]
-      sections -> Enum.map(sections, &String.trim/1)
-    end
-  end
-
-  defp extract_document_content(bytes, final_url, content_type, document_type, opts) do
-    case ExtractousEx.extract_from_bytes(bytes, opts[:extractous]) do
-      {:ok, %{content: content, metadata: metadata}} when is_binary(content) ->
-        {:ok, String.trim(content), normalize_metadata(metadata)}
-
-      {:error, reason} ->
-        {:error,
-         Error.adapter_error("ExtractousEx failed while extracting document content", %{
-           error_code: :unavailable,
-           url: final_url,
-           content_type: content_type,
-           document_type: document_type,
-           reason: reason
-         })}
-    end
-  rescue
-    error ->
-      {:error,
-       Error.adapter_error("ExtractousEx failed while extracting document content", %{
-         error_code: :unavailable,
-         url: final_url,
-         content_type: content_type,
-         document_type: document_type,
-         reason: error
-       })}
-  end
-
-  defp response_content_type(response) do
-    response.headers
-    |> response_header("content-type")
-    |> List.first()
-    |> case do
-      nil -> infer_content_type(response.body)
-      content_type -> content_type |> String.split(";") |> hd() |> String.trim() |> String.downcase()
-    end
-  end
-
-  defp response_header(headers, name) when is_map(headers) do
-    headers
-    |> Map.get(name, Map.get(headers, String.downcase(name), []))
-    |> List.wrap()
-  end
-
-  defp infer_content_type(body) when is_binary(body) do
-    cond do
-      String.starts_with?(body, "%PDF-") ->
-        "application/pdf"
-
-      likely_text?(body) ->
-        "text/plain"
-
-      true ->
-        "application/octet-stream"
-    end
-  end
-
-  defp infer_content_type(_body), do: "application/octet-stream"
-
-  defp text_content_type?(content_type) do
-    content_type in @text_content_types or String.starts_with?(content_type, "text/")
-  end
-
-  defp retrieved_at do
-    DateTime.utc_now()
-    |> DateTime.truncate(:second)
-    |> DateTime.to_iso8601()
-  end
-
-  defp estimate_tokens(content) when is_binary(content) do
-    div(String.length(content) + 3, 4)
-  end
-
-  defp estimate_tokens(_content), do: 0
-
-  defp extractable_document_type(content_type, final_url, body) do
-    Map.get(@document_content_types, content_type) ||
-      infer_document_type_from_body(body) ||
-      if(ambiguous_binary_content_type?(content_type), do: infer_document_type_from_url(final_url), else: nil)
-  end
-
-  defp infer_document_type_from_url(url) do
-    url
-    |> URI.parse()
-    |> Map.get(:path, "")
-    |> Path.extname()
-    |> String.trim_leading(".")
-    |> String.downcase()
-    |> case do
-      "" -> nil
-      extension -> Map.get(@document_extensions, extension)
-    end
-  end
-
-  defp infer_document_type_from_body(body) when is_binary(body) do
-    if String.starts_with?(body, "%PDF-"), do: :pdf, else: nil
-  end
-
-  defp infer_document_type_from_body(_body), do: nil
-
-  defp document_title(metadata, url) do
-    metadata
-    |> metadata_title()
-    |> blank_to_nil()
-    |> case do
-      nil -> title_from_url(url)
-      title -> title
-    end
-  end
-
-  defp metadata_title(metadata) when is_map(metadata) do
-    Enum.find_value([:title, "title", "dc:title", :"dc:title"], fn key ->
-      metadata
-      |> Map.get(key)
-      |> metadata_value_to_string()
-      |> blank_to_nil()
-    end)
-  end
-
-  defp metadata_value_to_string(nil), do: nil
-  defp metadata_value_to_string(value) when is_binary(value), do: String.trim(value)
-
-  defp metadata_value_to_string(value) when is_list(value),
-    do: value |> Enum.map_join(" ", &to_string/1) |> String.trim()
-
-  defp metadata_value_to_string(value) when is_atom(value), do: value |> Atom.to_string() |> String.trim()
-  defp metadata_value_to_string(value) when is_number(value), do: value |> to_string() |> String.trim()
-  defp metadata_value_to_string(_value), do: nil
-
-  defp normalize_metadata(metadata) when is_map(metadata), do: metadata
-
-  defp maybe_put_metadata(response, metadata) when metadata in [%{}, nil], do: response
-  defp maybe_put_metadata(response, metadata), do: Map.put(response, :metadata, metadata)
-
-  defp matching_section_indexes(sections, terms) do
-    downcased_terms = Enum.map(terms, &String.downcase/1)
-
-    sections
-    |> Enum.with_index()
-    |> Enum.flat_map(fn {section, index} ->
-      if section_matches_term?(section, downcased_terms), do: [index], else: []
-    end)
-  end
-
-  defp section_matches_term?(section, downcased_terms) do
-    lowered = String.downcase(section)
-    Enum.any?(downcased_terms, &String.contains?(lowered, &1))
-  end
-
-  defp expand_focus_window(matching_indexes, window, section_count) do
-    matching_indexes
-    |> Enum.flat_map(fn index -> (index - window)..(index + window) end)
-    |> Enum.filter(&(&1 >= 0 and &1 < section_count))
-    |> Enum.uniq()
-    |> Enum.sort()
-  end
-
-  defp render_section_slice(sections, indexes) do
-    indexes
-    |> Enum.map(&Enum.at(sections, &1))
-    |> Enum.reject(&(&1 == ""))
-    |> Enum.join("\n\n")
-    |> String.trim()
-  end
-
-  defp title_from_url(url) do
-    path = URI.parse(url).path || ""
-
-    case path do
-      "" -> nil
-      "/" -> nil
-      value -> value |> Path.basename() |> String.trim("/") |> blank_to_nil()
-    end
-  end
-
-  defp blank_to_nil(nil), do: nil
-  defp blank_to_nil(""), do: nil
-  defp blank_to_nil(value), do: value
-
-  defp ambiguous_binary_content_type?(content_type) do
-    content_type in [
-      "application/octet-stream",
-      "binary/octet-stream",
-      "application/download",
-      "application/x-download",
-      "application/zip",
-      "application/x-zip-compressed"
-    ]
-  end
-
-  defp likely_text?(body) when is_binary(body) do
-    String.valid?(body) and not String.contains?(body, <<0>>)
   end
 end
